+++ lib/hive_dart.dart
@@ -64,3 +64,4 @@ export 'src/util/number.dart';
 export 'src/util/random.dart';
 export 'src/util/string.dart';
 export 'src/util/time.dart';
+export 'src/util/url.dart';